+++ service/controller/chart/resource/status/types.go
@@ -1,0 +1,11 @@
+package status
+
+import v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
+type Request struct {
+	AppVersion   string  `json:"app_version"`
+	LastDeployed v1.Time `json:"last_deployed"`
+	Reason       string  `json:"reason"`
+	Status       string  `json:"status"`
+	Version      string  `json:"version"`
+}